requirements.txt
newsapi-python
langchain
sentence-transformers
faiss-cpu
openai
python-dotenv
